tambahkan verb filter untuk BoothController.php

+++ admin/controllers/BoothController.php
@@ -4,22 +4,36 @@ namespace admin\controllers;
 
 use common\models\Booth;
 use common\models\User;
+use Throwable;
 use Yii;
 use yii\data\ActiveDataProvider;
+use yii\filters\VerbFilter;
+use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\Response;
 
-class BoothController extends \yii\web\Controller
+class BoothController extends Controller
 {
-    public function behaviors()
+    /**
+     * @return array
+     */
+    public function behaviors(): array
     {
         return [
             'verbs'=>[
-                'class'=>'yii\filters\VerbFilter',
-
+                'class'=>VerbFilter::class,
+                'actions' => [
+                    'tolak'=>['POST'],
+                    'terima'=>['POST']
+                ]
             ]
-        ]
+        ];
     }
-    public function actionIndex()
+
+    /**
+     * @return string
+     */
+    public function actionIndex(): string
     {
         $data = Booth::find();
         $dataProvider = new ActiveDataProvider(['query' => $data]);
@@ -27,7 +41,10 @@ class BoothController extends \yii\web\Controller
         return $this->render('index', ['dataProvider' => $dataProvider]);
     }
 
-    public function actionVerifikasi()
+    /**
+     * @return string
+     */
+    public function actionVerifikasi(): string
     {
         $data = Booth::find()->where(['status' => Booth::STATUS_CREATED]);
         $dataProvider = new ActiveDataProvider(['query' => $data]);
@@ -35,13 +52,23 @@ class BoothController extends \yii\web\Controller
         return $this->render('verifikasi', ['dataProvider' => $dataProvider]);
     }
 
-    public function actionView($id)
+    /**
+     * @param $id
+     * @return string
+     * @throws NotFoundHttpException
+     */
+    public function actionView($id): string
     {
         $model = $this->findModel($id);
         return $this->render('view', compact('model'));
     }
 
-    function findModel($id)
+    /**
+     * @param $id
+     * @return Booth|null
+     * @throws NotFoundHttpException
+     */
+    protected function findModel($id)
     {
         $model = Booth::findOne($id);
         if (!$model) {
@@ -51,40 +78,70 @@ class BoothController extends \yii\web\Controller
         return $model;
     }
 
-    public function actionTerima($id)
+    /**
+     * @param $id
+     * @return Response
+     * @throws NotFoundHttpException
+     * @throws \Exception
+     */
+    public function actionTerima($id): Response
     {
         $model = $this->findModel($id);
-        $model->status = Booth::STATUS_VERIFIED;
-        $model->user->has_booth = User::HAS_BOOTH;
-        $model->user->save(false);
-        $model->save(false);
+        if (isset($model)) {
+            $model->status = Booth::STATUS_VERIFIED;
+            $model->user->has_booth = User::HAS_BOOTH;
+            $model->user->save(false);
+            $model->save(false);
 
-        $auth = Yii::$app->authManager;
-        $role = $auth->getRole('penjual');
-        $auth->assign($role, $model->user->getId());
+            $auth = Yii::$app->authManager;
+            $role = $auth->getRole('penjual');
+            $auth->assign($role, $model->user->getId());
 
 
-        Yii::$app->session->setFlash('success', 'Berhasil menyetujui verifikasi');
-        return $this->redirect(['booth/verifikasi']);
+            Yii::$app->session->setFlash('success', 'Berhasil menyetujui verifikasi');
+            return $this->redirect(['booth/verifikasi']);
+        }
+        throw new NotFoundHttpException();
     }
 
-    public function actionTolak($id)
+
+    /**
+     * @param $id
+     * @return Response
+     * @throws NotFoundHttpException
+     * @throws Throwable
+     * @throws yii\db\StaleObjectException
+     */
+    public function actionTolak($id): Response
     {
         $model = $this->findModel($id);
-        $model->user->has_booth = 0;
-        $model->user->update(false);
-        $model->delete();
-        Yii::$app->session->setFlash('success', 'Berhasil menolak verifikasi');
-
-        return $this->redirect(['index']);
+        if ($model !== null) {
+            $model->user->has_booth = 0;
+            $model->user->update(false);
+            $model->delete();
+            Yii::$app->session->setFlash('success', 'Berhasil menolak verifikasi');
+            return $this->redirect(['index']);
+        }
+        throw new NotFoundHttpException();
     }
 
-    public function actionDelete($id)
+    /**
+     * @param $id
+     * @return Response
+     * @throws NotFoundHttpException
+     * @throws Throwable
+     * @throws yii\db\StaleObjectException
+     */
+    public function actionDelete($id): Response
     {
         $model = $this->findModel($id);
-        $model->user->has_booth = 0;
-        $model->delete();
-        Yii::$app->session->setFlash('success', 'Berhasil menghapus Booth');
-        return $this->redirect(['index']);
+        if (isset($model)) {
+            $model->user->has_booth = 0;
+            $model->user->save(false);
+            $model->delete();
+            Yii::$app->session->setFlash('success', 'Berhasil menghapus Booth');
+            return $this->redirect(['index']);
+        }
+        throw new NotFoundHttpException();
     }
 }
